requirements.txt
crewai>=0.5.0
langchain>=0.1.0
openai>=1.6.1
python-dotenv==1.0.0
google-api-python-client==2.107.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
PyPDF2==3.0.1
pandas==2.1.1
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0